Refactor: Address coderabbit comments
