+++ rust/python_bindings/src/tests/mod.rs
@@ -2,43 +2,19 @@
 // Use of this source code is governed by the Apache License 2.0
 // that can be found in the LICENSE file.
 
-use std::sync::OnceLock;
-
 use pyo3::types::PyAnyMethods as _;
 
 use crate::bindings;
-use crate::schema_store::STORE;
 
 mod passwords;
 mod validation;
 
-static INIT_PY: OnceLock<()> = OnceLock::new();
-static INIT_STORE: OnceLock<()> = OnceLock::new();
-
 fn setup_python() {
-    INIT_PY.get_or_init(|| {
-        pyo3::append_to_inittab!(bindings);
-        pyo3::Python::initialize();
-    });
+    test_schema_store::setup_python(|| pyo3::append_to_inittab!(bindings));
 }
 
 fn setup() {
-    setup_python();
-    INIT_STORE.get_or_init(|| {
-        pyo3::Python::attach(|py| {
-            init_test_store(py);
-        });
-    });
-}
-
-fn init_test_store(py: pyo3::Python<'_>) {
-    assert!(STORE.get().is_none(), "Already set");
-    let module = py.import("_bindings").unwrap();
-    let args = ();
-    let kwargs = pyo3::types::PyDict::new(py);
-    let file = py.detach(test_schema_store::get_store_gz_path);
-    kwargs.set_item("file", file).unwrap();
-    let _ = module.call_method("init_store_from_file", args, Some(&kwargs));
+    test_schema_store::setup_python_with_store(|| pyo3::append_to_inittab!(bindings));
 }
 
 fn get_path_and_message_from_py_violation(
